chore: 增加了大量接口的注释，删除了inno setup中自带的一些sample
