--- InstallerGenerate/Inno Setup 6/Examples/64Bit.iss
@@ -1,34 +1,0 @@
-; -- 64Bit.iss --
-; Demonstrates installation of a program built for the x64 (a.k.a. AMD64)
-; architecture.
-; To successfully run this installation and the program it installs,
-; you must have a "x64" edition of Windows or Windows 11 on Arm.
-
-; SEE THE DOCUMENTATION FOR DETAILS ON CREATING .ISS SCRIPT FILES!
-
-[Setup]
-AppName=My Program
-AppVersion=1.5
-WizardStyle=modern
-DefaultDirName={autopf}\My Program
-DefaultGroupName=My Program
-UninstallDisplayIcon={app}\MyProg.exe
-Compression=lzma2
-SolidCompression=yes
-OutputDir=userdocs:Inno Setup Examples Output
-; "ArchitecturesAllowed=x64compatible" specifies that Setup cannot run
-; on anything but x64 and Windows 11 on Arm.
-ArchitecturesAllowed=x64compatible
-; "ArchitecturesInstallIn64BitMode=x64compatible" requests that the
-; install be done in "64-bit mode" on x64 or Windows 11 on Arm,
-; meaning it should use the native 64-bit Program Files directory and
-; the 64-bit view of the registry.
-ArchitecturesInstallIn64BitMode=x64compatible
-
-[Files]
-Source: "MyProg-x64.exe"; DestDir: "{app}"; DestName: "MyProg.exe"
-Source: "MyProg.chm"; DestDir: "{app}"
-Source: "Readme.txt"; DestDir: "{app}"; Flags: isreadme
-
-[Icons]
-Name: "{group}\My Program"; Filename: "{app}\MyProg.exe"
