--- InstallerGenerate/Inno Setup 6/Examples/Components.iss
@@ -1,34 +1,0 @@
-; -- Components.iss --
-; Demonstrates a components-based installation.
-
-; SEE THE DOCUMENTATION FOR DETAILS ON CREATING .ISS SCRIPT FILES!
-
-[Setup]
-AppName=My Program
-AppVersion=1.5
-WizardStyle=modern
-DefaultDirName={autopf}\My Program
-DefaultGroupName=My Program
-UninstallDisplayIcon={app}\MyProg.exe
-OutputDir=userdocs:Inno Setup Examples Output
-
-[Types]
-Name: "full"; Description: "Full installation"
-Name: "compact"; Description: "Compact installation"
-Name: "custom"; Description: "Custom installation"; Flags: iscustom
-
-[Components]
-Name: "program"; Description: "Program Files"; Types: full compact custom; Flags: fixed
-Name: "help"; Description: "Help File"; Types: full
-Name: "readme"; Description: "Readme File"; Types: full
-Name: "readme\en"; Description: "English"; Flags: exclusive
-Name: "readme\de"; Description: "German"; Flags: exclusive
-
-[Files]
-Source: "MyProg.exe"; DestDir: "{app}"; Components: program
-Source: "MyProg.chm"; DestDir: "{app}"; Components: help
-Source: "Readme.txt"; DestDir: "{app}"; Components: readme\en; Flags: isreadme
-Source: "Readme-German.txt"; DestName: "Liesmich.txt"; DestDir: "{app}"; Components: readme\de; Flags: isreadme
-
-[Icons]
-Name: "{group}\My Program"; Filename: "{app}\MyProg.exe"
