--- InstallerGenerate/Inno Setup 6/Examples/Example2.iss
@@ -1,27 +1,0 @@
-; -- Example2.iss --
-; Same as Example1.iss, but creates its icon in the Programs folder of the
-; Start Menu instead of in a subfolder, and also creates a desktop icon.
-
-; SEE THE DOCUMENTATION FOR DETAILS ON CREATING .ISS SCRIPT FILES!
-
-[Setup]
-AppName=My Program
-AppVersion=1.5
-WizardStyle=modern
-DefaultDirName={autopf}\My Program
-; Since no icons will be created in "{group}", we don't need the wizard
-; to ask for a Start Menu folder name:
-DisableProgramGroupPage=yes
-UninstallDisplayIcon={app}\MyProg.exe
-Compression=lzma2
-SolidCompression=yes
-OutputDir=userdocs:Inno Setup Examples Output
-
-[Files]
-Source: "MyProg.exe"; DestDir: "{app}"
-Source: "MyProg.chm"; DestDir: "{app}"
-Source: "Readme.txt"; DestDir: "{app}"; Flags: isreadme
-
-[Icons]
-Name: "{autoprograms}\My Program"; Filename: "{app}\MyProg.exe"
-Name: "{autodesktop}\My Program"; Filename: "{app}\MyProg.exe"
