--- InstallerGenerate/Inno Setup 6/Examples/ISPPExample1License.txt
@@ -1,4 +1,0 @@
-#pragma option -e+
-{#AppName} version {#AppVersion} License
-
-Bla bla bla
