--- InstallerGenerate/Inno Setup 6/Examples/License.txt
@@ -1,1 +1,0 @@
-This is the LICENSE file for My Program.
