--- InstallerGenerate/Inno Setup 6/Examples/MyDll/C#/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="UnmanagedExports" version="1.2.7" targetFramework="net45" />
-</packages>
