--- InstallerGenerate/Inno Setup 6/Examples/MyDll/C/MyDll.c
@@ -1,6 +1,0 @@
-#include <windows.h>
-
-void __stdcall MyDllFunc(HWND hWnd, char *lpText, char *lpCaption, UINT uType)
-{
-  MessageBox(hWnd, lpText, lpCaption, uType);
-}
