--- InstallerGenerate/Inno Setup 6/Examples/MyDll/C/MyDll.def
@@ -1,2 +1,0 @@
-EXPORTS
-  MyDllFunc
