--- InstallerGenerate/Inno Setup 6/Examples/MyDll/Delphi/MyDll.dpr
@@ -1,14 +1,0 @@
-library MyDll;
-
-uses
-  Windows;
-
-procedure MyDllFunc(hWnd: Integer; lpText, lpCaption: PAnsiChar; uType: Cardinal); stdcall;
-begin
-  MessageBoxA(hWnd, lpText, lpCaption, uType);
-end;
-
-exports MyDllFunc;
-
-begin
-end.
