--- InstallerGenerate/Inno Setup 6/Examples/Readme-Dutch.txt
@@ -1,1 +1,0 @@
-Dit is het Leesmij bestand voor My Program.
