--- InstallerGenerate/Inno Setup 6/Examples/Readme-German.txt
@@ -1,1 +1,0 @@
-Dies ist die LIESMICH-Datei für "My Program".
