--- InstallerGenerate/Inno Setup 6/Examples/Readme.txt
@@ -1,1 +1,0 @@
-This is the README file for My Program.
