--- InstallerGenerate/Inno Setup 6/Examples/UninstallCodeExample1.iss
@@ -1,46 +1,0 @@
-; -- UninstallCodeExample1.iss --
-;
-; This script shows various things you can achieve using a [Code] section for Uninstall.
-
-[Setup]
-AppName=My Program
-AppVersion=1.5
-WizardStyle=modern
-DefaultDirName={autopf}\My Program
-DefaultGroupName=My Program
-UninstallDisplayIcon={app}\MyProg.exe
-OutputDir=userdocs:Inno Setup Examples Output
-
-[Files]
-Source: "MyProg.exe"; DestDir: "{app}"
-Source: "MyProg.chm"; DestDir: "{app}"
-Source: "Readme.txt"; DestDir: "{app}"; Flags: isreadme
-
-[Code]
-function InitializeUninstall(): Boolean;
-begin
-  Result := MsgBox('InitializeUninstall:' #13#13 'Uninstall is initializing. Do you really want to start Uninstall?', mbConfirmation, MB_YESNO) = idYes;
-  if Result = False then
-    MsgBox('InitializeUninstall:' #13#13 'Ok, bye bye.', mbInformation, MB_OK);
-end;
-
-procedure DeinitializeUninstall();
-begin
-  MsgBox('DeinitializeUninstall:' #13#13 'Bye bye!', mbInformation, MB_OK);
-end;
-
-procedure CurUninstallStepChanged(CurUninstallStep: TUninstallStep);
-begin
-  case CurUninstallStep of
-    usUninstall:
-      begin
-        MsgBox('CurUninstallStepChanged:' #13#13 'Uninstall is about to start.', mbInformation, MB_OK)
-        // ...insert code to perform pre-uninstall tasks here...
-      end;
-    usPostUninstall:
-      begin
-        MsgBox('CurUninstallStepChanged:' #13#13 'Uninstall just finished.', mbInformation, MB_OK);
-        // ...insert code to perform post-uninstall tasks here...
-      end;
-  end;
-end;
